+++ modules/stac-server/opensearch_domain.tf
@@ -3,9 +3,9 @@ resource "random_id" "suffix" {
 }
 
 resource "aws_opensearch_domain" "stac_server_opensearch_domain" {
-  count           = var.deploy_stac_server_opensearch_serverless ? 0 : 1
-  domain_name     = lower(var.opensearch_stac_server_domain_name_override == null ? "${local.name_prefix}-stac-server" : var.opensearch_stac_server_domain_name_override)
-  engine_version  = var.opensearch_version
+  count          = var.deploy_stac_server_opensearch_serverless ? 0 : 1
+  domain_name    = lower(var.opensearch_stac_server_domain_name_override == null ? "${local.name_prefix}-stac-server" : var.opensearch_stac_server_domain_name_override)
+  engine_version = var.opensearch_version
 
   cluster_config {
     instance_type            = var.opensearch_cluster_instance_type
@@ -147,9 +147,9 @@ EOF
 
 resource "null_resource" "cleanup_opensearch_master_password_secret" {
   triggers = {
-    opensearch_master_password_secret     = "${local.name_prefix}-stac-server-master-creds-${random_id.suffix.hex}"
-    region                                = data.aws_region.current.name
-    account                               = data.aws_caller_identity.current.account_id
+    opensearch_master_password_secret = "${local.name_prefix}-stac-server-master-creds-${random_id.suffix.hex}"
+    region                            = data.aws_region.current.name
+    account                           = data.aws_caller_identity.current.account_id
   }
 
   provisioner "local-exec" {
@@ -210,9 +210,9 @@ EOF
 
 resource "null_resource" "cleanup_opensearch_stac_user_password_secret" {
   triggers = {
-    opensearch_stac_user_password_secret  = "${local.name_prefix}-stac-server-user-creds-${random_id.suffix.hex}"
-    region                                = data.aws_region.current.name
-    account                               = data.aws_caller_identity.current.account_id
+    opensearch_stac_user_password_secret = "${local.name_prefix}-stac-server-user-creds-${random_id.suffix.hex}"
+    region                               = data.aws_region.current.name
+    account                              = data.aws_caller_identity.current.account_id
   }
 
   provisioner "local-exec" {
@@ -287,27 +287,11 @@ resource "aws_lambda_function" "stac_server_opensearch_user_initializer" {
   ]
 }
 
-resource "null_resource" "invoke_stac_server_opensearch_user_initializer" {
-  count     = var.deploy_stac_server_opensearch_serverless ? 0 : 1
-  triggers  = {
-    INITIALIZER_LAMBDA                 = aws_lambda_function.stac_server_opensearch_user_initializer.function_name
-    OPENSEARCH_HOST                    = var.deploy_stac_server_opensearch_serverless ? aws_opensearchserverless_collection.stac_server_opensearch_serverless_collection[0].collection_endpoint : aws_opensearch_domain.stac_server_opensearch_domain[0].endpoint
-    OPENSEARCH_MASTER_CREDS_SECRET_ARN = aws_secretsmanager_secret.opensearch_master_password_secret.arn
-    OPENSEARCH_USER_CREDS_SECRET_ARN   = aws_secretsmanager_secret.opensearch_stac_user_password_secret.arn
-    REGION                             = data.aws_region.current.name
-  }
+resource "aws_lambda_invocation" "invoke_stac_server_opensearch_user_initializer" {
+  count         = var.deploy_stac_server_opensearch_serverless ? 0 : 1
+  function_name = aws_lambda_function.stac_server_opensearch_user_initializer.function_name
 
-  provisioner "local-exec" {
-    interpreter = ["bash", "-ec"]
-    command     = <<EOF
-export AWS_DEFAULT_REGION=${data.aws_region.current.name}
-export AWS_REGION=${data.aws_region.current.name}
-
-echo "Creating stac_server user on OpenSearch cluster."
-aws lambda invoke --function-name ${aws_lambda_function.stac_server_opensearch_user_initializer.function_name} --payload '{ }' output
-
-EOF
-  }
+  input = "{}"
 
   depends_on = [
     random_password.opensearch_master_password,
@@ -317,5 +301,18 @@ EOF
     aws_secretsmanager_secret.opensearch_stac_user_password_secret,
     aws_secretsmanager_secret_version.opensearch_stac_user_password_secret_version,
     aws_lambda_function.stac_server_opensearch_user_initializer
+  ]
+}
+
+resource "aws_lambda_invocation" "stac_server_opensearch_domain_ingest_create_indices" {
+  count         = var.deploy_stac_server_opensearch_serverless ? 0 : 1
+  function_name = aws_lambda_function.stac_server_ingest.function_name
+
+  input = "{ \"create_indices\": true }"
+
+  depends_on = [
+    aws_lambda_function.stac_server_ingest,
+    aws_lambda_invocation.invoke_stac_server_opensearch_user_initializer,
+    aws_opensearch_domain.stac_server_opensearch_domain
   ]
 }