+++ profiles/analytics/credentials/admin/inputs.tf
@@ -1,4 +1,4 @@
 variable "credentials_name_prefix" {
-  description   = "The name prefix for admin credentials."
-  type          = string
+  description = "The name prefix for admin credentials."
+  type        = string
 }