+++ profiles/analytics/credentials/admin/main.tf
@@ -20,14 +20,14 @@ resource "aws_secretsmanager_secret" "filmdrop_analytics_credentials" {
 }
 
 resource "aws_secretsmanager_secret_version" "filmdrop_analytics_credentials_version" {
-  secret_id     = aws_secretsmanager_secret.filmdrop_analytics_credentials.id
+  secret_id = aws_secretsmanager_secret.filmdrop_analytics_credentials.id
   secret_string = jsonencode({
-    USERNAME  = "admin"
-    PASSWORD  = random_password.filmdrop_analytics_credentials.result
+    USERNAME = "admin"
+    PASSWORD = random_password.filmdrop_analytics_credentials.result
   })
 
   lifecycle {
-      ignore_changes = [secret_string, secret_binary]
+    ignore_changes = [secret_string, secret_binary]
   }
 }
 
@@ -36,13 +36,13 @@ resource "aws_secretsmanager_secret" "filmdrop_analytics_dask_secret_token" {
 }
 
 resource "aws_secretsmanager_secret_version" "filmdrop_analytics_dask_secret_token_version" {
-  secret_id     = aws_secretsmanager_secret.filmdrop_analytics_dask_secret_token.id
+  secret_id = aws_secretsmanager_secret.filmdrop_analytics_dask_secret_token.id
   secret_string = jsonencode({
-    PROXYTOKEN  = random_password.filmdrop_analytics_dask_secret_token_proxy_token.result
-    APITOKEN    = random_password.filmdrop_analytics_dask_secret_token_api_token.result
+    PROXYTOKEN = random_password.filmdrop_analytics_dask_secret_token_proxy_token.result
+    APITOKEN   = random_password.filmdrop_analytics_dask_secret_token_api_token.result
   })
 
   lifecycle {
-      ignore_changes = [secret_string, secret_binary]
+    ignore_changes = [secret_string, secret_binary]
   }
 }