+++ profiles/analytics/credentials/main.tf
@@ -1,6 +1,6 @@
 module "admin_creds" {
-    source  = "./admin"
-    count   = var.create_credentials ? 1 : 0
+  source = "./admin"
+  count  = var.create_credentials ? 1 : 0
 
-    credentials_name_prefix = var.credentials_name_prefix
+  credentials_name_prefix = var.credentials_name_prefix
 }